Reorganize Components into their own folders and update imports
