+++ src/Components/App/App.js
@@ -1,13 +1,14 @@
 import React, { Component } from "react";
-import MoviesContainer from "./MoviesContainer";
-import Header from './Header'
-import Poster from './Poster';
+import MoviesContainer from "../MoviesContainer/MoviesContainer";
+import Header from '../Header/Header'
+import Poster from '../Poster/Poster';
 import PropTypes from 'prop-types';
-import Error from './Error'
-import movieData from '../Data/data';
-import {fetchMovieCollection} from './apiCalls'
+import Error from '../Error/Error'
+import movieData from '../../Data/data';
+import {fetchMovieCollection} from '../apiCalls'
 import { Route, Switch } from 'react-router-dom';
 import './App.css';
+
 class App extends Component {
   constructor() {
     super();
